[API] return EmailingAttempt in mailer
