+++ api/lib/domain/usecases/publish-session.js
@@ -38,19 +38,15 @@ async function _sendPrescriberEmails(session) {
 
   const emailingAttempts = [];
   for (const recipientEmail of recipientEmails) {
-    try {
-      await mailService.sendCertificationResultEmail({
-        email: recipientEmail,
-        sessionId: session.id,
-        sessionDate: session.date,
-        certificationCenterName: session.certificationCenter,
-        resultRecipientEmail: recipientEmail,
-        daysBeforeExpiration: 30,
-      });
-      emailingAttempts.push(EmailingAttempt.success(recipientEmail));
-    } catch (error) {
-      emailingAttempts.push(EmailingAttempt.failure(recipientEmail));
-    }
+    const emailingAttempt = await mailService.sendCertificationResultEmail({
+      email: recipientEmail,
+      sessionId: session.id,
+      sessionDate: session.date,
+      certificationCenterName: session.certificationCenter,
+      resultRecipientEmail: recipientEmail,
+      daysBeforeExpiration: 30,
+    });
+    emailingAttempts.push(emailingAttempt);
   }
   return emailingAttempts;
 }
@@ -60,34 +56,6 @@ function _distinctCandidatesResultRecipientEmails(certificationCandidates) {
     .map((candidate) => candidate.resultRecipientEmail)
     .filter(Boolean);
 }
-
-class EmailingAttempt {
-  constructor(recipientEmail, status) {
-    this.recipientEmail = recipientEmail;
-    this.status = status;
-  }
-
-  hasFailed() {
-    return this.status === AttemptStatus.FAILURE;
-  }
-
-  hasSucceeded() {
-    return this.status === AttemptStatus.SUCCESS;
-  }
-
-  static success(recipientEmail) {
-    return new EmailingAttempt(recipientEmail, AttemptStatus.SUCCESS);
-  }
-
-  static failure(recipientEmail) {
-    return new EmailingAttempt(recipientEmail, AttemptStatus.FAILURE);
-  }
-}
-
-const AttemptStatus = {
-  SUCCESS: 'SUCCESS',
-  FAILURE: 'FAILURE',
-};
 
 function _someHaveSucceeded(emailingAttempts) {
   return some(emailingAttempts, (emailAttempt) => emailAttempt.hasSucceeded());
